chore: add tests for nonexistent file checks
Added a collection of [contrived] tests for exercising the "nonexistent
files in build output" check logic. Note that the string values asserted
in these tests originate from `package-builder/validate-build.bash`.
Required adding a new set of mock data for two packages, one of which
is in the "toplevel" package group and another that was not, and we
didn't have one of those yet.