refactor: drop internal config caching
We previously implemented full config caching,
so that multiple calls to `Config::parse` won't incur repeated I/O to the config files.

However, this caused worse tests that call `Config::parse`
to share global state potentially breaking test isolation.
Hence we in `#[cfg(test)]` environments, the cache was disabled.

Further, the cache caused `ensure_environment_trust` to update the config file
on disk but fail to correctly update the in-memory config via a `Config::parse`.

Since we only parse the config once[^1] at the beginning of the `flox` process
and subsequently consume the parsed `Config` object,
the caching wasn't even ever effective in a positive way.

This commit drops the `OnceLock` caching in favor of always performing a fresh read.

[^1]: technically twice but i suspect we can consolidate that.