fix: collect narinfo and send with storepath
There were 2 issues preventing the correct narinfo from getting to the
catalog for nix-copy stores.

The first was that the "output" field was
being populated with the output "name" and not the "storepath", so when
status was asked for the storepath, it was never found.

The second one I suspect was a output format change in nix path-info,
causing the deserialization of the narinfo to come back completely
empty.  That is fixed and a test was added.