fix(publish): Improve remote branch heuristics
Simplify the logic and fix two bugs..

1. Relaxes the constraint so that we no longer return an error if your
branch is tracking a remote but you've pushed that branch to a different
remote. This can happen if you've cloned a project, committed changes to
the default branch, then forked it, and pushed to your fork. Previously
we would only check the tracked remote and now we go on to check the
others.

2. Tightens the constraint so that we now return an error if you have
more than one remote and the current rev isn't pushed to any of them.
Previously we would return the first regardless and now we check that
they do have the rev.

The remotes are still checked in the same overall order as before and
each one is only checked once.