fix: invoke `version.command` string from activated environment
Issue #3308 documents the issue of attempting to use a command installed
from the manifest in the process of using the `version.command` attribute
to set the version string for a build, and finding that the command is
not found.

This patch updates the code to always invoke the command string from within
an invocation of `$(FLOX_ENV)/activate -c '<string>'`, which brings the
full context of the activation into scope for running the command. It also
adds an additional check to the existing suite of version checks to exercise
this use case.