fix: silence attach retry warning
Silence the attach retry warning until retries are exhausted.

Running two `flox activate -- cmd` in a row can lead to one activation
trying to attach, the other one exiting, and then the first having to
start rather than attach.

The retry warning is causing CI to flake, and it isn't actually helpful
to users. See
https://github.com/flox/flox/actions/runs/12714454841/job/35444689929#step:5:352
```
Prior activation of the environment completed before it could be attached to.
Retrying ...
```

Silence the warning since it's not helpful.

I considered holding the lock for the first attach attempt, but special
casing the first iteration in wait_for_activation_ready_and_attach_pid
seems like it adds more complexity than it's worth to save us 100
milliseconds when we hit a somewhat rare edge case.