feat(publish): Add progress spinner for repo state
This function and its children call a bunch of `git` commands, some of
which also contact remotes over the network, so provide some feedback to
the user about why this might take a second or two.

Instrumenting here, rather than the child functions, because each
individual command might not take very long and would result in very
short spinners.