fix: wrap manifest build executibles with correct bash
Commit 30437d2 provided a workaround for the fact that the build
wrapper environment provides one version of a bash, but then in the
process of performing a manifest build we were wrapping executables
using a version of bash corresponding to the nixpkgs version used
for the build, which was not always a known/valid path to appear
in the output closure. The workaround was to shell out to Nix to
derive the version of bash corresponding to the nixpkgs commit
being used, but that had the effect of [potentially] adding another
copy of bash to the closure.

This fix takes a different approach:

1. add `libexec/build-wrapper-bash` symlink to the build wrapper
   environment pointing to the version of bash used when building it
2. override the `makeWrapper` hook used when performing the manifest
   build to use the above as the `@shell@` value to be substituted
   rather than the default as derived from the nixpkgs revision used
   for the build.