refactor: allow checking git status for nix expression targets
Attempting expression builds without a containing git repository is prevented
and replied with a detailed error message with the previous commit.

However, just initializing a git repo, let that check pass
causing errors later within the build subsystem, which are comparatively crude:

```
/Users/dcarley/projects/flox/flox/build/flox-package-builder/libexec/flox-build.mk:155: *** no manifest or Nix expression builds found in /nix/store/2fx8ai0jgzv1r2j2ijxhgj1xxpbhqwnj-environment-develop.  Stop.
❌ ERROR: Build failed with status: exit status: 2
```

The reason is that `fetchTree` as used by the by NEF causes all _untracked_ files
to be filtered out, including the expressions we attempt to build.
This commit ensures that the nix files defining the expressions to be built
are in fact tracked, and won't cause errors (as easily) up the stack.

Note that nix may still fail evaluating or building expressions
because required _source_ files or nix imports are unavailable.