fix: use cpio for serializing directory contents
Issue #3310 documents problems encountered using gnutar and sed to
serialize and modify directory contents in the process of copying the
manifest build output from a temporary directory to $out. This approach
had the effect of modifying the destination of symbolic links as embedded
in the gnutar control stream, which in turn corrupted its own validation
checksums.

Fortunately, cpio archives do not suffer from this issue and are tolerant
of binary substitutions that have the effect of rewriting symlink
destinations, and this commit switches from the use of tar to cpio for
copying and modifying manifest build outputs.