fix: evaluate build-manifest.nix using nixpkgs URL from flox-interpreter
We've now employed a few different workarounds to use the "right" nixpkgs
revision when wrapping executables in manifest builds, but our attempts
have been limited to replacing `bash` and `sh` only and did not provide
the assurance that patchShebangs could replace other "interpreters" with
ones from the correct nixpkgs revision.

There are only two possible nixpkgs revisions for use in a manifest build:

1. the revision used to build the `flox-interpreter` package
  - will be the same used to build `flox` itself
  - guaranteed to be defined
  - guaranteed to be recent enough to contain all dependencies required
    by `build-manifest.nix` (e.g. `t3`)
2. the revision used to build all packages in the "toplevel" package group
  - will not be defined in an environment containing no packages
  - may not be recent enough to satisfy the requirements of `build-manifest.nix`

This patch addresses the problem once and for all by using the first of these
when evaluating `build-manifest.nix`. It does this by embedding a copy of the
nixpkgs URL used to build the `flox-interpreter` within its package outputs,
and then updates `build-manifest.nix` to use that flakeref when populating its
own `pkgs` variable.

Also updated the existing tests as required and added another to ensure that
`patchShebangs` can pull in arbitrary coreutil paths from outside `$FLOX_ENV`
and still get one derived from the correct nixpkgs revision.

Further background and discussion can be found in: #3327