fix: do not relay stdout/err coming from builder
Unlike Nix builds which solely write to stderr, the manifest builder
emits lines to both stdout and stderr, which presents the possibility
of getting the ordering wrong when printing those lines to the tty.
We wrote `t3` specifically to address this issue while at the same
time providing color highlighting and timestamping, but then undid
all that by again reading and writing each line of output from flox,
thereby re-introducing the possibility that lines can come out of
order.

This patch updates `FloxBuildMk::{clean,build}` to inherit stdio
rather than relaying output and streaming possibly out of order
at the CLI level.
For the use in tests, or other deliberate use where the output
should be suppressed and colelcted instead,
`FloxBuildMk::new_with_buffers`, was added to provide `&mut String`
references that will be filled with the the output instead.