refactor: use common bundled nixpkgs for all build evaluations
Nix expression builds, specifically extend an arbitrary nixpkgs set
and are therefore passed a dynamically determined instance of nixpkgs.
All other evaluations involving nixpkgs should use a common nixpkgs bundled with flox.
This ensures that
a) for a given version of flox we know which nixpkgs it uses for evaluations and build
b) have test coverage for the invariant pieces involving nixpkgs
c) dont need to implement backwards compatibility of our manifest-build.nix script
   to accommodate changes in nixpkgs (e.g. `bash[[Non]Interactive`, `t3`).
d) in the context of the preceeding commits, have a stronger guarantee that prefetching
   the common nixpkgs flake (or straight bundling of it) provides all required nixpkgs
   for manifest builds.

Semantically this doesn't change much;
We currently build the environment which carries a reference to the nixpkgs
shipped with the _building flox_, prior to a (package) build
and therefore indirectly ensure that we reuse common nixpkgs in practice.
The only change here is that we dont carry that information through the environment build,
which might also allow to avoid environment rebuilds in the future if necessary.