feat(generations): Expose meta w/new env trait
Add a new environment trait for generations and expose metadata for
managed and remote environments.

Currently this returns `GenerationsError` rather than wrapping it in
something like `EnvironmentError`. We might want to change that in the
future, but right now it's hard to predict without wiring up all of the
functionality.