feat(init): Add push hint when not in git
So that people can discover the benefits of using FloxHub when creating
new environments. With the exception of environments that are within
existing git repos - detailed in the code comment.

The check is performed early in the `init` command so that traverse
directories to find the git parent doesn't add a noticeable delay
between printing parts of the result message. It shouldn't take long
enough to warrant a spinner though.

The original criteria was to print this notice on the first invocation
of `activate` for an environment and to also exclude temporary
environments (within `$TMPDIR` or name `tmp.*`) but that would have
required adding detection and state tracking to `flox activate` which is
much more sensitive to latency.

We don't need the temporary environment logic in this implementation
because it's part of an existing message that can be suppressed with
`--quiet` for automation.

The additional `message::plain("")` is necessary to preserve the
trailing newline that we had before, now that we don't know which
message will be last. Unfortunately it seems that Bats can't assert
whether it's there or not though.