feat(push): Normalise successful push message
Refactor into a single function that only differs the heading message
depending on whether the environment is newly created or not. The
indication that it's public is folded into the heading when first
creating an environment because it doesn't seem worth using a whole line
to describe this.

The rest of the next step actions can be the same. We now promote
FloxHub foremost, then the ability to activate the remote without
pulling, and then pulling to make a "copy".

I don't think it's worth distinguishing between `flox pull` and `flox
pull --copy` here like FloxHub does given that we have limited line
width and attention span.