fix: mount /private/tmp in podman VM
Previously we mounted `/tmp` because this is where `bats` creates
temp directories for test runs. This is necessary for files from tests
to be available in the VM.

The new Podman 5.5.0 release prevents you from mounting `/tmp`, so
we can instead mount `/private/tmp` on macOS since `/tmp` is an alias
anyway. See this Podman issue for details:
https://github.com/containers/podman/issues/18230