feat(envs): Show URLs for managed and remote
So that it's easier to discover FloxHub functionality. This also
disambiguates the names of remote environments by including their owner.

There's an existing "bug" whereby remote environments that you're
currently or have previously used are listed with their
`~/.cache/flox/remote` directories and they aren't omitted if you
currently have one of them activated due to the way that we cache them
as managed environments.

I'm not attempting to fix that here because I think the fix belongs in
how we model the environments rather than applying fixes to the display
layer.