refactor: use local services to generate mocks
The idea here is to use local versions of FloxHub/`catalog-server` when
recording mocks from unit tests. This means that we need to start those
services and populate all relevant config prior to recording mocks.

This requires that the FloxHub services are running in a separate shell.
I first attempted to activate the FloxHub environment inline, but the
FloxHub environment sources the output of `nix print-dev-env` in its
`hook.on-activate` script, which removes directories from `PATH` needed
by the Flox devshell.

This also gets a Nixpkgs revision from the base catalog for populating
publish metadata.

There are some authentication-related variables that are extracted from
the FloxHub environment as well since we can run the data-generation
process from inside the FloxHub environment.