test: ensure communication with local server
This adds a test that ensures we can get base catalog information from
the local catalog-server, and does some necessary refactoring along the
way.

I don't think this is the *most* important test, but it required no real
setup of a project, a passing build, etc and allowed me to ensure that
communication and authentication was set up properly for the local
services.