fix: don't pass --already-sourced-env-dirs flag at bootstrap
Modern versions of tcsh support the ":Q" modifier for passing empty args
on the command line, but versions prior to 6.23 do not have a way to do
that, so to support these versions we will instead avoid passing the
--already-sourced-env-dirs argument altogether when there is no default
value to be passed. This will in turn require updating flox-activations
to provide a default [empty] value for `already_sourced_env_dirs` when
not provided/overridden by CLI args.