refactor: derive `Copy` for `GenerationId`
`GenerationId` is just a container for an usize and already implements `Clone`.
As a typed accessor index it might be debated if it should indeed implement any of these,
or rather be lifetime-bound to the metadata object.
It's probably not worth the additional complexity right now though, so for now we improve the ergonomics around the ability to copy the id.