test(usage): Remove regex
This regex was matching anything because the `|` weren't escaped and
allowed it to perform a partial match on any one of the sub-commands.

This also appeared to mask an issue with it not supporting the `\d`
shorthand character class.

I don't love the `--partial` matches to ignore the first line of the
output, because it can mask other unexpected output, but it's much safer
than using regexes.