fix: patch Nix for better single-user mode error
I have an upstream PR applying this patch, but it hasn't merged yet, and
even if it does we can apply it in the meantime.

When running as a non-root user using a single-user installation of Nix,
or when the nix daemon crashes, the current error Nix provides doesn't
give much hint for how to solve the issue. The current error for a
`flox install` is:
```
❌ ERROR: Failed to build environment:

Failed to realise 'hello':
error: opening lock file '/nix/var/nix/db/big-lock': Permission denied
```

Prior to our most recent update of Nix, the error included

```
error: could not set permissions on '/nix/var/nix/profiles/per-user' to 755: Operation not permitted
```

Modify the error to provide some hints at how to solve the problem.
After this patch, the error is:
```
❌ ERROR: Failed to build environment:

Failed to realise 'hello':
error:
       … This command may have been run as non-root in a single-user Nix installation,
       or the Nix daemon may have crashed.

       error: opening lock file '/nix/var/nix/db/big-lock': Permission denied
```