test(usage): Fix race condition
This passed locally after fixing the regex in a preciding commit and
when running the `just unit-tests usage.bats` but failed in CI when run
with all of the other tests because other files had registered
environments and triggered a different conditional.

Fix it by changing the assertion message and ensuring that we have at
least one inactive environment. I've copied `project_setup` so that we
init the project in a test scoped directory.