fix: allow building when git tracked files are deleted
To avoid build failures due to deleted tracked files as described in the
previous commit, we either need improve the error messaging, or permit
missing tracked files and build with only the tracked files present.
Requiring less git fumbling in the way to builds seems preferable from a
UX perspective, so that's what this Commit addresses.

Because 'git ls-files' does not have a flag to filter deleted files, but
allows to _only_ show deleted files, use 'comm' to do the filtering for
us.

* drop `wait $!` after `tar` command
We dont perform process substitution anymore but read from a single file.
waiting for `$!` fails with

```
/nix/store/nqh4lv0w7n7cw7n8vvn2gfzv804y4hk6-bash-interactive-5.2p37/bin/bash: line 1: wait: `': not a pid or valid job spec
```