fix: race conditions in publish tests
This addresses race conditions in the publish tests. The race condition
looks like this:
- All of the tests currently use the same package name.
- Any two tests that operate on the same package and the same catalog
  will see different states.

For most tests we don't want to test that we can create the catalog, we
want to test operations that take place after the catalog has been
created. For this reason we'll create two different users:
- A user whose personal catalog has not yet been created.
- A user whose personal catalog has already been created (by the time
  we run tests with that user).

This allows us to use the first user to ensure that a personal catalog
is created inline with a user's first publish, while using the second
user to not worry about creating their personal catalog.

By using separate package names for each test, we also ensure that there
are no collisions between different tests.