feat(activate): zsh compinit logs in verbose
Log the reason for regenerating completion dumps for tracelevels greater
or equal to `flox activate -v`. This will make it easier to debug where
time is being spent in activation.

I don't think it's worth the overhead of testing different zsh versions
for this change, so here's proof of the difference:

    % echo $ZSH_VERSION
    5.9
    % compinit -w -d /dev/null
    No existing compdump file found, regenerating

    % echo $ZSH_VERSION
    5.8.1
    % compinit -w -d /dev/null
    % echo $?
    0