fix(install): Prevent '=' and ' ' in pkg-path
Prevent malformed pkg-paths from `flox install` commands from causing
exceptions in `catalog-server` by checking for characters that should
never appear after parsing out the `@version` requirement.

This exception:

- https://flox-dev.sentry.io/issues/6777203580/?project=4507306107273216&referrer=github_integration

Appears to have been caused by the equivalent (values changed for
anonymity but structure preserved) of:

    flox install 'foo.pkg_group = "bar_group"'

Which could either be an attempt to imperatively install to a group, a
copy/paste mistake from a manifest, or an LLM agent.

We have to be relatively flexible about the strings that we accept here
because, unlike the typed manifest, the `@version` specifier can contain
equals and whitespace. However the `pkg-path` never should.