feat: impl `HistorySpec::summary()`
Dynamically produce a short summary of a single change.
This summary is intended to capture the _intent_ when the change was made
and should be cheap to compute.
It therefore isnt blocked on any sort of diff'ing
of the underlying environment.

* refactor: use structured kind information

In order to dynamically generate consistent summaries,
and in preparation of better auditability of history items,
we drop the `summary: String` field, and add `target` attributes for
relevant `HistoryKind` variants.