Revert "fix(install): Prevent '=' and ' ' in pkg-path"
This reverts commit 0ae5c95b4275a6dd347281f5fc8e59fbea1cc095.

We now have a server-side fix in `catalog-server` which prevents badly
formed pkg/attr paths from causing errors. They are now treated as
literal strings which don't match any known package:

    % FLOX_CATALOG_URL=https://api.preview.flox.dev/ nix run github:flox/flox/v1.6.0 -- install -d ~/tmp 'foo.pkg_group = "bar_group"'
    ❌ ERROR: resolution failed:
    Could not find package 'foo.pkg_group = "bar_group"'.
    Try 'flox search' with a broader search term.

Implemented in:

- https://github.com/flox/floxhub/pull/303

To be deployed to production in:

- https://github.com/flox/floxhub/pull/305

CLI validation would still be useful if we could do it consistently in
`manifest::typed` to handle both imperative and declarative installs,
but since we can't do that because serde untagged enums swallow any
deserialization errors from further down the stack, I think it's better
not to special case validation for imperative installs and instead just
rely on `catalog-server`.