fix(buildenv): Don't outdent null script fields
Don't call `outdentScript` on `null` fields as rendered by older
versions of Flox. This doubles up the `getAttr` call but I think that's
preferable to more nested conditionals.

I was unable to pull this older environment with Flox versions v1.5.0 and
later, but pulls with v1.4.2 and earlier still worked, and so do
`activate -r` presumably because it hadn't need to rebuild:

    % nix run github:flox/flox/v1.5.0 -- pull dcarley/tailscale
    ❌ ERROR: Failed to build environment:

    Failed to constructed environment: error:
           … while calling the 'derivationStrict' builtin
             at <nix/derivation-internal.nix>:37:12:
               36|
               37|   strict = derivationStrict drvAttrs;
                 |            ^
               38|

           … while evaluating derivation 'environment'
             whose name attribute is located at /nix/store/ffh0da5n98rmmrsqd1l3z1janlck84l3-flox-buildenv-0.0.1/lib/buildenv.nix:252:11

           … while evaluating attribute 'exportReferencesGraph' of derivation 'environment'
             at /nix/store/ffh0da5n98rmmrsqd1l3z1janlck84l3-flox-buildenv-0.0.1/lib/buildenv.nix:325:3:
              324|   # the contents of requisites.txt for each output.
              325|   exportReferencesGraph.graph = inputSrcs ++ [
                 |   ^
              326|     interpreter_out

           (stack trace truncated; use '--show-trace' to show the full, detailed trace)

           error: expected a string but found null: null

I couldn't figure out exactly which version of Flox that environment was
last locked with and when we stopped rendering those specific fields, so
I just generated the lockfile manually.

I don't think that `build` is technically susceptible to the same
problem or worth testing because `build.command` isn't optional.