test(services): Increase redis timeout
This has been failing frequently in CI on `x86_64-darwin` which we know
has issues with resource contention and slow builds.

    not ok 470 can start redis-server and access it using redis-cli # in 19537 ms
    # tags: services
    # (from function `assert_success' in file /nix/store/rq4i8l8hn5i3lvb1byx7qr5bmi6dcgi2-bats-with-libraries-1.12.0/share/bats/bats-assert/src/assert_success.bash, line 42,
    #  in test file services.bats, line 160)
    #   `assert_success' failed
    # ✨ Created environment 'test' (x86_64-darwin)
    #
    # Next:
    #   $ flox search <package>    <- Search for a package
    #   $ flox install <package>   <- Install a package into an environment
    #   $ flox activate            <- Enter the environment
    #   $ flox edit                <- Add environment variables and shell hooks
    #   $ flox push                <- Use the environment from other machines or
    #                                 share it with someone on FloxHub
    # ✅ Environment successfully updated.
    #
    # -- command failed --
    # status : 124
    # output :
    # --

Set a timeout on each `redis-cli` invocation because the default
(AFAICT) means that it can wait forever with no output if the socket is
bound but not responding, which it could be due to the way that we use
`portgrab`. This flag wasn't in the version of Redis that we had so I
had to upgrade it by deleting the mock and running `just mk-data`:

> -t <timeout>       Server connection timeout in seconds (decimals allowed).
>                    Default timeout is 0, meaning no limit, depending on the OS.

Then increase the total timeout to allow for 5 attempts of 1s + 0.1s.