feat(generations): provide better forward compatibility support
For `AllGenerationsMetadata` and `HistorySpec`, *permit* unknown fields
and deserialize them into a separate map.

ideally we would like this to look like

```rust
#[derive(Deserialize,...)]
struct Foo {
    field_1: String,
    field_2: usize,
    ...
    #[serde(flatten)]
    other: Map<String, Value>
}
```
However, there are several issues in serde preventing this,
e.g. when mixed with another `#[serde(flatten)]`ed field [1,2].

As a workaround we borrow the solution from alloy-rs [3],
i.e. wrap structs that require forward compatibility for added fields
in a common wrapper interface `WithOtherFields<T>`,
which deserializes both the original type and the additional fields,
taking care to _correctly_ handle deserialization of the additionals.

We add `Deref[Mut]` implementations to make this new wrapper
as transparent as possible, providing access to the known fields only.

[1]: https://github.com/serde-rs/serde/issues/1909
[2]: https://github.com/serde-rs/serde/issues/2200
[3]: https://github.com/alloy-rs/alloy/pull/495