refactor: use flox-activations for zsh fpath
This change refactors our Zsh init script to use `flox-activations` to
set `fpath`/`FPATH` in a similar way that we use `flox-activations` to
set `PATH`, `MANPATH`, and other `PATH`-like variables.

The main thing to notice here is that since `flox-activations` operates
on colon-separated variables already, we take `FPATH` as an input, but
generate code that sets `fpath`. Setting `FPATH` directly has unexpected
behavior that causes completions installed to the system to be removed
from `fpath`/`FPATH`.