docs(manifest): Update link in template
The redirect without the trailing slash is broken.

We'll need to fix this for existing manifests that have been generated
by `flox init`, but it's not trivial right now, and we can avoid the
indirection for newer versions:

- https://github.com/flox/floxdocs/issues/307

We moved the manifest concept to the environments concept for reasons
described in the following PR, which I still think make sense:

- https://github.com/flox/floxdocs/pull/226

But if you want to learn more about the manifest then it's probably
better that you go to the reference page for that specifically.