fix(generations): Don't increment on noops
We shouldn't create new generations and history for actions that are
noops, such as:

- `install` the same package
- `upgrade` when there are no changes
- `include upgrade` includes when there are no changes

The following already worked correctly:

- `edit` with no changes
- `uninstall` a package that isn't present

I've added tests for all of these except `include upgrade` because it's
fiddly to orchestrate and I want to get this fix into the v1.7 release.

The tests are a bit boilerplatey and could be improved in the future but
I think they provide value in testing the integration of generations
with environments and preventing regressions.

There are also large discrepancies between how we determine whether
there are changes from each of the results which would be better served
by enums but I'm not going to fix now for the same reason.