fix: wording tweak: current -> live
Use "live" instead of "current" when referring to generations.

Also print "(live)" instead of "(current)" in `flox generations list`

The word "current" means "most recent" which has an implication about
time.

Because you can switch to an older generation that is not "most recent,"
"current" has some incorrect implications. We avoid the word "latest"
for the same reasons (although with "latest," it's a bit more overt)

Because of it's implications about time, "current" is also hard to use
in documentation or messaging about the past. "Previously current" or
"last current" are oxymorons.