fix(generations): Cast Flox.argv to Option later
Fixes a test failure where `edit` passed an empty vec and generations
attempted to replace `argv[0]` with `"flox"` on an invalid index. This
happens for tests and would happen for FloxHub changes in the future.

This still doesn't quite feel right because `HistorySpec` could
theoretically (but not practically right now) contain `Some(vec![])`. I
originally went the other way and pushed the `Option` further up but
that made that problem worse.