fix: omit "stubs" output in buildenv.nix
This is a temporary fix to omit the `stubs` output of packages that
install stub libraries to be used at build time, when other libraries
are needed at run time.

This is a mitigation for an issue caused by our recursive linking:
```
lib-stubs-pkg/
  lib/
    mylib.so <-- really a stub
    stubs/
      mylib.so <-- really a stub
lib-pkg/
  lib/
    otherlib.so
    stubs --> lib-stubs-pkg/lib/stubs
FLOX_ENV/
  lib/
    mylib.so <-- really a stub
    otherlib.so
    stubs/
      ...
```

When `lib-stubs-pkg` and `lib-pkg` are merged, the stub library
`lib-stubs-pkg/lib/mylib.so` is brought in at `FLOX_ENV/lib/mylib.so`,
putting it on the default search path for libraries. This is fine at
build time, but causes the binary to try to use the stub library
at runtime (in the same environment).

Since the `libs` output of the package contains a symlink to the
`stubs` output, it knows how to reach the stubs if necessary.