fix: omit "stubs" outputs in Perl builder
After omitting the `stubs` output in `buildenv.nix`, the stub libraries
were still found in the recursively linked environment. We tracked this
down to the Perl builder recursively including propagated build inputs.

This change omits *all* outputs whose store paths end in `-stubs`, which
is how a store path would be formatted for a `stubs` output of a larger
package. We could tailor this to be more specific.