refactor: make change producing platform extensible
`ChangeSpec::author` and `CahngeSpec::hostname` are currently required.
As we speak of changes made via floxhub, they will either
not be relevant or have different semantics,
depending on where the change was made.
However, we would be forced to use the same names in our current schema.
This commit proposes to extensibly represent different platforms,
i.e. local changes and soon floxhub changes, so that their respective
metadata can be represented without being forced to a common schema.