fix(watchdog): Remove expired activation PIDs
This ensures that we don't make unnecessary system calls to check PIDs
and write logs when PIDs have previously been observed as terminated.

I noticed whilst looking into an activation and services bug that the
previous fixes in 879bf96b and 0d43c6a7 weren't working because I was
still getting logs of log entries, even after upgrading and restarting
the activations with a new watchdog:

    1.5G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.83a5fa64.log.2025-08-15
    1.5G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.83a5fa64.log.2025-08-16
    1.5G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.83a5fa64.log.2025-08-17
    1.1G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.83a5fa64.log.2025-08-18
    2.6G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.df661826.log.2025-08-15
    2.6G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.df661826.log.2025-08-16
    2.6G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.df661826.log.2025-08-17
    1.6G	/Users/dcarley/dotfiles/flox/default/.flox/log/watchdog.df661826.log.2025-08-18

There were three causes for this, which I equally missed in code review:

1. The condition `self.pids_watching != old_pids` was only ever
   triggered on the first iteration of the loop or when new activations
   were added because it was always comparing a list of PIDs that had
   already been pruned of terminations.
2. When it was triggered we wrote back an unmodified version of the
   `activations.json` because `self.update_watchlist` only pruned the
   internal state of `self.pids_watching` and returned the original
   `maybe_locked_activations` that was read from the file.
3. The test didn't catch this because it only asserted that the running
   `pid2` was still present and not that the stopped `pid2` had been
   removed.

Ultimately `update_watchlist` was hard to reason about in terms of state
and locking so I've repurposed it to:

1. No longer keep a separate internal state.
2. Return `WaitResult` directly when there are no PIDs left.
3. Update `activations.json` if there are PIDs removed.

This obviously increases the amount of time that we hold the lock in
each polling loop because we have to check all of the PIDs first. We
previously talked about not locking on reads in order to reduce
filesystem calls within the tight loop but I no longer think that this
is possible because we have to hold the lock whilst checking and
deciding whether we need to write back, otherwise we will wipe out any
activations that are simultaneously added at the same time. I'm pretty
sure that the biggest improvement to performance will be not checking or
logging already terminated PIDs though.

The test has been updated to assert on the entire contents of the PIDs
rather than making individual assertions about them, which makes it
slightly less likely to make mistakes and easier to debug when there are
failures.

The logic and tests for removing terminated PIDs has been moved to
`Activations`, which feels like a better fit alongside the functionality
for modifying a singular activation by adding and removing PIDs, and in
the same crate as `proc_status`.

I've also manually validated this by activating an environment in two
terminals, tailing the watchdog logs, and exiting one activation. We now
only print one log entry for the first time that the PID is seen as
terminated and it's no longer visible in:

    jq . $(dirname $_FLOX_ACTIVATION_STATE_DIR)/activations.json