test(activate): Modify activations.json after
This failed with the following error after modifying the watchdog to
write terminated PIDs back to `activations.json` because it now needs to
call `check_version()` and the modification within an existing
activation causes the watchdog to exit prematurely:

    # -- command failed --
    # status : 1
    # output (46 lines):
    #   Error: This environment has already been activated with an incompatible version of 'flox'.
    #
    #   Exit all activations of the environment and try again.
    #   PIDs of the running activations: 69334

Fix this by modifying the version after the activation has exited. I'm
not sure why I chose to do it within the activation in the first place
since it doesn't present what would happen in real life. I can only
assume that it was a simple copy/paste from the preceding test which
does need to modify in-place.